Attribute Macro debris_derive::object

#[object]

Handles an impl block containing methods for debris objects

Usage:

#[object(Type::Foo)]
impl FooStruct {
    // mark functions for the object with #[method] or #[special]
    #[special]
    fn add(_ctx: &CompileContext, a: &FooStruct, b: &FooStruct) -> FooStruct {
        (a.value + b.value).into()
    }

    // If the function can fail (at compile time), return a [LangResult]:
    #[method]
    fn might_fail(ctx: &CompileContext, value: &FooStuct) -> LangResult<FooStruct> {
        Ok(0.into())
    }

    // The `CompileContext` can also be taken by &mut reference
    #[method]
    fn foo(ctx: &mut CompileContext) -> FooStruct {
        ctx.emit(...);
        Ok(0.into())
    }
     
    // Overloads are also supported
    #[special]
    fn add(_ctx: &CompileContext, argument: &FooStruct) -> FooStruct {...}
}

The user could now run this code (note that a is of type Foo): let b = a.might_fail(other_a);

Note that overloads must have different arguments (which is not checked right now).

Implementation

This proc-macro collects all functions marked as #[special] or #[method] and removes them from the impl body. Then, this macro implementes the trait HasClass which returns the associated class of this struct. This class contains all marked methods as ObjectFunction which can be called from debris code.

Since this operation is quite expensive, it is only run once and then cached.